move model parsing to utils
