+++ lib/seg_face.py
@@ -10,6 +10,7 @@ current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentfra
 parent_dir = os.path.dirname(current_dir)
 sys.path.insert(0, parent_dir) 
 from multithreading import BackgroundGenerator
+from utils import parse_model_weights
 
 class Mask():
     def __init__(self):
@@ -324,141 +325,5 @@ class Mask():
         preds = np.array(Q, dtype=np.float32).reshape((n_classes, h, w)).transpose(1, 2, 0)
         return preds
 
-        
-    def parse_model_weights(self, weight_file, model_file):
-    
-        # handles grouped convolutions
-        def convolution(weights_dict, name, input, group, conv_type, filters=None, **kwargs):
-            if not conv_type.startswith('layer'):
-                layer = keras.applications.mobilenet.DepthwiseConv2D(name=name,
-                                                                     **kwargs)(input)
-                return layer
-                
-            if group == 1:
-                func = getattr(layers, conv_type.split('.')[-1])
-                layer = func(name = name, filters = filters, **kwargs)(input)
-                return layer
-                
-            group_list = []
-            weight_groups = []
-            if not weights_dict == None:
-                w = numpy.array(weights_dict[name]['weights'])
-                weight_groups = numpy.split(w, indices_or_sections=group, axis=-1)
-                
-            grouped_channels = int(filters / group)
-            for c in range(group):
-                x = layers.Lambda(lambda z: z[:, :, :, c * grouped_channels:(c + 1) * grouped_channels])(input)
-                x = layers.Conv2D(name=name + "_" + str(c), filters=grouped_channels, **kwargs)(x)
-                weights_dict[name + "_" + str(c)] = dict()
-                weights_dict[name + "_" + str(c)]['weights'] = weight_groups[c]
-                group_list.append(x)
-                
-            layer = layers.concatenate(group_list, axis = -1)
-            
-            if 'bias' in weights_dict[name]:
-                b = K.variable(weights_dict[name]['bias'], name = name + "_bias")
-                layer = layer + b
-                
-            return layer
-            
-        def load_weights_from_file(weight_file):
-            try:
-                weights_dict = numpy.load(weight_file).item()
-            except:
-                weights_dict = numpy.load(weight_file, encoding='bytes').item()
-                
-            return weights_dict
-            
-        def set_layer_weights(model, weights_dict):
-            for layer in model.layers:
-                if layer.name in weights_dict:
-                    cur_dict = weights_dict[layer.name]
-                    current_layer_parameters = []
-                    
-                    if layer.__class__.__name__ == "BatchNormalization":
-                        if 'scale' in cur_dict:
-                            current_layer_parameters.append(cur_dict['scale'])
-                        if 'bias' in cur_dict:
-                            current_layer_parameters.append(cur_dict['bias'])
-                        current_layer_parameters.extend([cur_dict['mean'],
-                                                        cur_dict['var']])
-                                                        
-                    elif layer.__class__.__name__ == "Scale":
-                        if 'scale' in cur_dict:
-                            current_layer_parameters.append(cur_dict['scale'])
-                        if 'bias' in cur_dict:
-                            current_layer_parameters.append(cur_dict['bias'])
-                            
-                    elif layer.__class__.__name__ == "SeparableConv2D":
-                        current_layer_parameters = [cur_dict['depthwise_filter'],
-                                                    cur_dict['pointwise_filter']]
-                        if 'bias' in cur_dict:
-                            current_layer_parameters.append(cur_dict['bias'])
-                            
-                    else:
-                        current_layer_parameters = [cur_dict['weights']]
-                        if 'bias' in cur_dict:
-                            current_layer_parameters.append(cur_dict['bias'])
-                            
-                    model.get_layer(layer.name).set_weights(current_layer_parameters)
-                    
-            return model
-
-        weights_dict = load_weights_from_file(weight_file)
-        
-        input        = layers.Input(name = 'input', shape = (300, 300, 3,) )
-        input_c      = layers.ZeroPadding2D(padding = ((100, 100), (100, 100)))(input)
-        conv1_1      = convolution(weights_dict, name='conv1_1', input=input_c, group=1, conv_type='layers.Conv2D', filters=64, kernel_size=(3, 3), activation='relu')
-        conv1_2      = convolution(weights_dict, name='conv1_2', input=conv1_1, group=1, conv_type='layers.Conv2D', filters=64, kernel_size=(3, 3), activation='relu', padding='same')
-        pool1        = layers.MaxPooling2D(name = 'pool1', pool_size = (2, 2), strides = (2, 2), padding='same')(conv1_2)
-        
-        conv2_1      = convolution(weights_dict, name='conv2_1', input=pool1, group=1, conv_type='layers.Conv2D', filters=128, kernel_size=(3, 3), activation='relu', padding='same')
-        conv2_2      = convolution(weights_dict, name='conv2_2', input=conv2_1, group=1, conv_type='layers.Conv2D', filters=128, kernel_size=(3, 3), activation='relu', padding='same')
-        pool2        = layers.MaxPooling2D(name = 'pool2', pool_size = (2, 2), strides = (2, 2), padding='same')(conv2_2)
-        
-        conv3_1      = convolution(weights_dict, name='conv3_1', input=pool2, group=1, conv_type='layers.Conv2D', filters=256, kernel_size=(3, 3), activation='relu', padding='same')
-        conv3_2      = convolution(weights_dict, name='conv3_2', input=conv3_1, group=1, conv_type='layers.Conv2D', filters=256, kernel_size=(3, 3), activation='relu', padding='same')
-        conv3_3      = convolution(weights_dict, name='conv3_3', input=conv3_2, group=1, conv_type='layers.Conv2D', filters=256, kernel_size=(3, 3), activation='relu', padding='same')
-        pool3        = layers.MaxPooling2D(name = 'pool3', pool_size = (2, 2), strides = (2, 2), padding='same')(conv3_3)
-        
-        conv4_1      = convolution(weights_dict, name='conv4_1', input=pool3, group=1, conv_type='layers.Conv2D', filters=512, kernel_size=(3, 3), activation='relu', padding='same')
-        conv4_2      = convolution(weights_dict, name='conv4_2', input=conv4_1, group=1, conv_type='layers.Conv2D', filters=512, kernel_size=(3, 3), activation='relu', padding='same')
-        conv4_3      = convolution(weights_dict, name='conv4_3', input=conv4_2, group=1, conv_type='layers.Conv2D', filters=512, kernel_size=(3, 3), activation='relu', padding='same')
-        pool4        = layers.MaxPooling2D(name = 'pool4', pool_size = (2, 2), strides = (2, 2), padding='same')(conv4_3)
-        
-        conv5_1      = convolution(weights_dict, name='conv5_1', input=pool4, group=1, conv_type='layers.Conv2D', filters=512, kernel_size=(3, 3), activation='relu', padding='same')
-        conv5_2      = convolution(weights_dict, name='conv5_2', input=conv5_1, group=1, conv_type='layers.Conv2D', filters=512, kernel_size=(3, 3), activation='relu', padding='same')
-        conv5_3      = convolution(weights_dict, name='conv5_3', input=conv5_2, group=1, conv_type='layers.Conv2D', filters=512, kernel_size=(3, 3), activation='relu', padding='same')
-        pool5        = layers.MaxPooling2D(name = 'pool5', pool_size = (2, 2), strides = (2, 2), padding='same')(conv5_3)
-        
-        fc6          = convolution(weights_dict, name='fc6', input=pool5, group=1, conv_type='layers.Conv2D', filters=4096, kernel_size=(7, 7), activation='relu')
-        drop6        = layers.Dropout(name = 'drop6', rate = 0.5, seed = None)(fc6)
-        fc7          = convolution(weights_dict, name='fc7', input=drop6, group=1, conv_type='layers.Conv2D', filters=4096, kernel_size=(1, 1), activation='relu')
-        drop7        = layers.Dropout(name = 'drop7', rate = 0.5, seed = None)(fc7)
-        
-        scale_pool3  = layers.Lambda(lambda x: x * 0.0001, name='scale_pool3')(pool3)
-        scale_pool4  = layers.Lambda(lambda x: x * 0.01, name='scale_pool4')(pool4)
-        score_pool3_r  = convolution(weights_dict, name='score_pool3_r', input=scale_pool3, group=1, conv_type='layers.Conv2D', filters=2, kernel_size=(1, 1))
-        score_pool4_r  = convolution(weights_dict, name='score_pool4_r', input=scale_pool4, group=1, conv_type='layers.Conv2D', filters=2, kernel_size=(1, 1))
-        score_pool3c = layers.Cropping2D(cropping=((9, 8), (9, 8)), name='score_pool3c')(score_pool3_r)
-        score_pool4c = layers.Cropping2D(cropping=((5, 5), (5, 5)), name='score_pool4c')(score_pool4_r)
-        
-        score_fr_r     = convolution(weights_dict, name='score_fr_r', input=drop7, group=1, conv_type='layers.Conv2D', filters=2, kernel_size=(1, 1))
-        upscore2_r   = convolution(weights_dict, name='upscore2_r', input=score_fr_r, group=1, conv_type='layers.Conv2DTranspose', filters=2, kernel_size=(4, 4), strides=(2, 2), use_bias=False)
-        fuse_pool4   = layers.add(name = 'fuse_pool4', inputs = [upscore2_r, score_pool4c])
-        
-        upscore_pool4_r= convolution(weights_dict, name='upscore_pool4_r', input=fuse_pool4, group=1, conv_type='layers.Conv2DTranspose', filters=2, kernel_size=(4, 4), strides=(2, 2), use_bias=False)
-        
-        fuse_pool3   = layers.add(name = 'fuse_pool3', inputs = [upscore_pool4_r, score_pool3c])
-        upscore8_r     = convolution(weights_dict, name='upscore8_r', input=fuse_pool3, group=1, conv_type='layers.Conv2DTranspose', filters=2, kernel_size=(16, 16), strides=(8, 8), use_bias=False)
-        score        = layers.Cropping2D(cropping=((31, 45), (31, 45)), name='score')(upscore8_r)
-        
-        model        = Model(inputs = [input], outputs = [score], name = 'face_seg_fcn_vgg16')
-        
-        set_layer_weights(model, weights_dict)
-        model.save(model_file)
-        
-        return model
-        
 if __name__ == '__main__':
     Mask()
